requirements.txt
# python version 3.8.17
grpcio-tools == 1.44.0
grpcio == 1.59.0
protobuf == 3.19.6
loguru == 0.7.2
pyunit_prime==2020.3.22
gmpy2
phe==1.5.0
scikit-learn